convert sessionID into accessToken
Updated the Tidal API wrapper to retrieve the accessToken once a sessionID is established. This means that users no longer need to use a packet sniffer or reverse proxy to find their access token. The refreshToken is stored, but not currently implemented.

The details for the API endpoints came from this source: https://github.com/openTIDAL/docTIDAL/wiki/Authentication